refactor: move calendar utils to calendar folder
