+++ src/components/_taskbar/Events/components/EventsHeader/EventsHeader.tsx
@@ -6,7 +6,7 @@ import { EventData } from 'types/store/plansState.type';
 
 import { TODAY_ID } from 'src/constants';
 
-import changeSelectedDateToDayName from 'utils/changeSelectedDateToDayName';
+import changeSelectedDateToDayName from 'utils/calendar/changeSelectedDateToDayName';
 
 import { ReactComponent as CloseIcon } from 'assets/icons/close.svg';
