+++ src/components/_taskbar/TimeAndDate/components/CurrentDate/CurrentDate.tsx
@@ -1,6 +1,6 @@
 import { useAppSelector } from 'store/hooks';
 
-import formatCurrentDate from 'utils/formatCurrentDate';
+import formatCurrentDate from 'utils/calendar/formatCurrentDate';
 
 import classes from './CurrentDate.module.css';
