+++ src/utils/calendar/getSplittedToday.ts
